final change/commit for now. various problems fixed
- added better user input error handling in vector_creations.rs and main.rs. the only one that doesn't have robust error handling is mode_answer() (and by extension,
mode_checker()) because it seemed annoying to have to check whether or not the user had all the right inputs when transferring values to a vector. it seemed easier to just
accept the invalid inputs and check the answers at the end in mode_checker(), even if the end result is a little messy and imperfect.

- deprecated the unneeded user input fns that returned a number in user_inputs.rs. having a more robust error handling means just accepting most inputs as a string, and then
later down the line check if the input is an acceptable number.

- changed the README to reflect the state of the program.

- made some minor changes to overview.txt

note for posterity: i was going to rewrite some of the program in idiomatic rust by using more iterator functions, but i looked and the program was already written pretty
idiomatically. i was also going to rearrange the files so all except main.rs would live inside a 'lib' folder, and main.rs would be cut down and the extra stuff would be
in a separate file, but i ultimately saw no need to do those things.